Changed func to utils.py
